Rename and replace first table view the new, second one.
